Change np.float to float

+++ museval/metrics.py
@@ -551,7 +551,7 @@ def _compute_projection_filters(G, sf, estimated_source):
     filters_len-1
     """
     # epsilon
-    eps = np.finfo(np.float).eps
+    eps = np.finfo(float).eps
 
     # shapes
     (nsampl, nchan) = estimated_source.shape
